Test Runner class for running separate testscopes
