+++ src/TestScope.js
@@ -12,98 +12,11 @@ class ComponentNotFoundError extends Error {
 
 export default class TestScope {
 
-  constructor(component, waitTime, startDelay, shouldSendReport) {
+  constructor(component, waitTime) {
     this.component = component;
     this.testHooks = component.testHookStore;
-
     this.testCases = [];
-
     this.waitTime = waitTime;
-    this.startDelay = startDelay;
-    this.shouldSendReport = shouldSendReport;
-
-    this.run.bind(this);
-  }
-
-  // Internal: Start tests after optional delay time.
-  async run() {
-    if (this.startDelay) {
-      await this.pause(this.startDelay);
-    }
-    this.runTests();
-  }
-
-  // Internal: Synchronously run each test case one after the other, outputting
-  // on the console if the test case passes or fails, and adding to testResult
-  // array for reporting purposes.
-  // Resets the app after each test case by changing the component key to force
-  // React to re-render the entire component tree.
-  async runTests() {
-    let testResults = [];
-    let errorCount = 0;
-
-    const start = new Date();
-    console.log(`Cavy test suite started at ${start}.`);
-
-    for (let i = 0; i < this.testCases.length; i++) {
-      let {description, f} = this.testCases[i];
-      try {
-        await f.call(this);
-        let successMsg = `${description}  ✅`;
-
-        console.log(successMsg);
-        testResults.push({message: successMsg, passed: true});
-      } catch (e) {
-        let errorMsg = `${description}  ❌\n   ${e.message}`;
-
-        console.warn(errorMsg);
-        testResults.push({message: errorMsg, passed: false});
-        errorCount += 1;
-      }
-      await this.component.clearAsync();
-      this.component.reRender();
-    }
-
-    const stop = new Date();
-    const duration = (stop - start) / 1000;
-    console.log(`Cavy test suite stopped at ${stop}, duration: ${duration} seconds.`);
-
-    const report = {
-      results: testResults,
-      errorCount: errorCount,
-      duration: duration
-    }
-
-    if (this.shouldSendReport) {
-      await this.sendReport(report);
-    }
-  };
-
-  sendReport(report) {
-    const url = 'http://127.0.0.1:8082/report';
-    const options = {
-      method: 'POST',
-      body: JSON.stringify(report),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    };
-
-    return fetch(url, options)
-      .then((response) => {
-        console.log('Cavy test report successfully sent to cavy-cli');
-      })
-      .catch((error) => {
-        if (error.message.match(/Network request failed/)) {
-          console.group(`Cavy test report server is not running at ${url}`);
-          console.log("If you are using cavy-cli, maybe it's not set up correctly or not reachable from this device?");
-          console.groupEnd();
-        } else {
-          console.group('Error sending test results')
-          console.warn(error.message);
-          console.groupEnd();
-        }
-      });
   }
 
   // Public: Find a component by its test hook identifier. Waits
